tower placement and free cursor
made player character be able to place and undo towers. you place towers by pressing E, and undo towers by pressing G. when holding E you get a free cursor to place your towers and the tower mesh is set to false as to not mess with other object and the player view. mesh becomes true again when action is released